One more issue with existing code, is that we use `gl.UNSIGNED_BYTE`, so index buffer might only of a `Uint8Array` which fits numbers up to `255`, so if object will have more than 255 vertices – it will be rendered incorrectly. Let's fix this

+++ src/gl-helpers.js
@@ -108,6 +108,6 @@ export function parseObj(objSource) {
 
     return { 
         vertices: new Float32Array(vertices), 
-        indices: new Uint8Array(indices),
+        indices: new Uint16Array(indices),
     };
 }
